fix sccp_device when using AO_compare_and_swap casting from int back to char pointer.
git-svn-id: https://svn.code.sf.net/p/chan-sccp-b/code/trunk@3530 43c1dc8c-776e-4df2-8ef1-6e829299be21

+++ src/sccp_device.c
@@ -2107,7 +2107,7 @@ void sccp_device_addMessageToStack(sccp_device_t * device, const uint8_t priorit
 #ifdef SCCP_BUILTIN_CAS_PTR
 		yieldedValue = __sync_val_compare_and_swap(&device->messageStack[priority], oldValue, newValue);		// Atomic Swap In newmsgptr
 #elif SCCP_ATOMIC_OPS
-		yieldedValue = (char *)((unsigned int)AO_compare_and_swap((size_t *)&device->messageStack[priority], (size_t)oldValue, (size_t)newValue));			// Atomic Swap using boemc atomic_ops library
+		yieldedValue = (char *)((size_t)AO_compare_and_swap((size_t *)&device->messageStack[priority], (size_t)oldValue, (size_t)newValue));			// Atomic Swap using boemc atomic_ops library
 #else
 		// would require lock
        		#warning "Atomic functions not implemented, please install libatomic_ops package to remedy. Otherwise problems are to be expected."
@@ -2143,7 +2143,7 @@ void sccp_device_clearMessageFromStack(sccp_device_t * device, const uint8_t pri
 #ifdef SCCP_BUILTIN_CAS_PTR
 		yieldedValue = __sync_val_compare_and_swap(&device->messageStack[priority], oldValue, newValue);		// Atomic Swap In newmsgptr
 #elif SCCP_ATOMIC_OPS
-		yieldedValue = (char *)((unsigned int)AO_compare_and_swap((size_t *)&device->messageStack[priority], (size_t)oldValue, (size_t)newValue));			// Atomic Swap using boemc atomic_ops library
+		yieldedValue = (char *)((size_t)AO_compare_and_swap((size_t *)&device->messageStack[priority], (size_t)oldValue, (size_t)newValue));			// Atomic Swap using boemc atomic_ops library
 #else
 		// would require lock
        		#warning "Atomic functions not implemented, please install libatomic_ops package to remedy. Otherwise problems are to be expected."
